Remove macOS and Windows from CI
Removed because the CI script has to install a full version of VISA from
NI/R&S/etc, which takes a long time.

I'm also unsure about whether that is allowed by their terms and
condition. I think it is, as their installer is not distributed (just
downloaded and used for testing).